ref: Add BaseInteractiveChart and BaseInteractiveChartState
